Implemented a non-parametric first stage
for probability of entry. I divide market variables into quintiles and
estimated entry probabilities by taking the means of markets with the
same quantiles. It's a very sketchy estimator because we have 5X5X5 =
125 possibilities but only 100 markets. Ciliberto and Tamer supplement
did the same, though: they had 2,000 markets but at least 12X12X12
groups O.o. I guess the only thing that's important is that it's
consistent.